requirements.txt
Django==3.2.14
django-ratelimit==3.0.1
uWSGI
requests
djangorestframework
drf-yasg
